requirements.txt
aiohttp==3.6.2
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==3.0.1
djangorestframework==3.11.0
Flask==1.1.1
idna==2.8
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
multidict==4.7.1
pkg-resources==0.0.0
pyee==6.0.0
python-decouple==3.3
pytz==2019.3
requests==2.22.0
six==1.13.0
slackclient==2.5.0
slackeventsapi==2.1.0
sqlparse==0.3.0
typing-extensions==3.7.4.1
urllib3==1.25.7
Werkzeug==0.16.0
yarl==1.4.2
